:art: Refactoring: move Middlewares to a separate file
